changed threshold distance argument from double to float in Line2DData::getPos
git-svn-id: https://svn.opendtect.org/od/trunk@27705 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiVis/uivispickretriever.cc
@@ -119,7 +119,7 @@ void uiVisPickRetriever::pickCB( CallBacker* cb )
 	pickedgeomid_ = s2dd->getGeomID();
 	PosInfo::Line2DPos pos2d;
 	const bool res =
-	    s2dd->getGeometry().getPos( pickedpos_, pos2d, mUdf(double) );
+	    s2dd->getGeometry().getPos( pickedpos_, pos2d, mUdf(float) );
 	if ( res )
 	    pickedtrcnr_ = pos2d.nr_;
     }
